tripletex-agent/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
openai>=1.0.0
httpx==0.27.0
pdfplumber==0.11.0
python-multipart==0.0.9
python-dotenv>=1.0.0
